--- pypeit/data/tests/output/keck_nires_A/keck_nires_A.pypeit
@@ -1,45 +1,0 @@
-# Auto-generated PypeIt input file using PypeIt version: 1.16.1.dev653+g73abd3b61
-# UTC 2025-07-03T22:51:44.524+00:00
-
-# User-defined execution parameters
-[rdx]
-    spectrograph = keck_nires
-
-# Setup
-setup read
-Setup A:
-  dispname: spec
-setup end
-
-# Data block 
-data read
- path /Users/olga/PypeIt-development-suite/RAW_DATA/keck_nires/ABBA_nostandard
-              filename |        frametype |               ra |               dec |       target | dispname |    decker | binning |              mjd |          airmass | exptime | dithpat | dithpos | dithoff | frameno | calib | comb_id | bkg_id
-NR.20201128.22795.fits | arc,science,tilt | 25.8722004481703 | -1.77957177216147 | eris_28 0620 |     spec | 0.55 slit |     1,1 |  59181.263841269 | 1.13659692763023 |   300.0 |    ABBA |       A |     5.0 |      49 |     0 |       1 |      2
-NR.20201128.23132.fits | arc,science,tilt | 25.8702924043135 | -1.77752676502195 | eris_28 0620 |     spec | 0.55 slit |     1,1 | 59181.2677371023 | 1.12734707860924 |   300.0 |    ABBA |       B |    -5.0 |      50 |     0 |       2 |      1
-NR.20201128.23467.fits | arc,science,tilt | 25.8702636585777 |  -1.7775294833213 | eris_28 0620 |     spec | 0.55 slit |     1,1 | 59181.2716109449 | 1.11888914544008 |   300.0 |    ABBA |       B |    -5.0 |      51 |     0 |       2 |      1
-NR.20201128.23803.fits | arc,science,tilt | 25.8720347222654 | -1.77964895855717 | eris_28 0620 |     spec | 0.55 slit |     1,1 | 59181.2754992551 | 1.11131755069033 |   300.0 |    ABBA |       A |     5.0 |      52 |     0 |       1 |      2
-NR.20201128.27090.fits | arc,science,tilt | 25.8713175226439 | -1.77935397652216 | eris_28 0720 |     spec | 0.55 slit |     1,1 | 59181.3135426579 | 1.07442213859665 |   300.0 |    ABBA |       A |     5.0 |      78 |     1 |       9 |     10
-NR.20201128.27429.fits | arc,science,tilt | 25.8696729456124 | -1.77709533959956 | eris_28 0720 |     spec | 0.55 slit |     1,1 | 59181.3174680051 | 1.07424479902789 |   300.0 |    ABBA |       B |    -5.0 |      79 |     1 |      10 |      9
-NR.20201128.27764.fits | arc,science,tilt | 25.8696731576493 |  -1.7770784323125 | eris_28 0720 |     spec | 0.55 slit |     1,1 | 59181.3213461301 | 1.07471568798519 |   300.0 |    ABBA |       B |    -5.0 |      80 |     1 |      10 |      9
-NR.20201128.28103.fits | arc,science,tilt | 25.8712014513636 | -1.77938177739209 | eris_28 0720 |     spec | 0.55 slit |     1,1 | 59181.3252714773 | 1.07582574161077 |   300.0 |    ABBA |       A |     5.0 |      81 |     1 |       9 |     10
-NR.20201128.28811.fits | arc,science,tilt | 25.8708414027903 | -1.77951938593208 | eris_28 0720 |     spec | 0.55 slit |     1,1 | 59181.3334704357 | 1.08039391687922 |   300.0 |    ABBA |       A |     5.0 |      82 |     1 |      13 |     14
-NR.20201128.29148.fits | arc,science,tilt | 25.8692550474395 | -1.77722150028739 | eris_28 0720 |     spec | 0.55 slit |     1,1 |  59181.337366269 | 1.08358124735467 |   300.0 |    ABBA |       B |    -5.0 |      83 |     1 |      14 |     13
-NR.20201128.29483.fits | arc,science,tilt | 25.8692263017036 | -1.77722420309764 | eris_28 0720 |     spec | 0.55 slit |     1,1 | 59181.3412384912 | 1.08737283214936 |   300.0 |    ABBA |       B |    -5.0 |      84 |     1 |      14 |     13
-NR.20201128.29819.fits | arc,science,tilt | 25.8708128623596 | -1.77948829452396 | eris_28 0720 |     spec | 0.55 slit |     1,1 | 59181.3451284218 | 1.09192101789746 |   300.0 |    ABBA |       A |     5.0 |      85 |     1 |      13 |     14
-NR.20201128.30168.fits | arc,science,tilt | 25.8707247988554 | -1.77953034939269 | eris_28 0720 |     spec | 0.55 slit |     1,1 | 59181.3491777273 | 1.09745765534104 |   300.0 |    ABBA |       A |     5.0 |      86 |     1 |      17 |     18
-NR.20201128.30505.fits | arc,science,tilt | 25.8691384427909 | -1.77723246422437 | eris_28 0720 |     spec | 0.55 slit |     1,1 | 59181.3530735606 | 1.10347880143997 |   300.0 |    ABBA |       B |    -5.0 |      87 |     1 |      18 |     17
-NR.20201128.30840.fits | arc,science,tilt | 25.8691096970552 | -1.77723516703461 | eris_28 0720 |     spec | 0.55 slit |     1,1 | 59181.3569457829 | 1.11021366511624 |   300.0 |    ABBA |       B |    -5.0 |      88 |     1 |      18 |     17
-NR.20201128.31179.fits | arc,science,tilt | 25.8706379923562 | -1.77953851114172 | eris_28 0720 |     spec | 0.55 slit |     1,1 | 59181.3608711301 | 1.11770611490234 |   300.0 |    ABBA |       A |     5.0 |      89 |     1 |      17 |     18
-NR.20201128.07158.fits |     lampoffflats |             58.0 |              45.0 |      unknown |     spec | 0.55 slit |     1,1 | 59181.0828553894 | 1.41291034446565 |   100.0 |    NONE |    None |     0.0 |      21 |   all |      -1 |     -1
-NR.20201128.07382.fits |     lampoffflats |             58.0 |              45.0 |      unknown |     spec | 0.55 slit |     1,1 | 59181.0854467088 | 1.41291034446565 |   100.0 |    NONE |    None |     0.0 |      22 |   all |      -1 |     -1
-NR.20201128.07496.fits |     lampoffflats |             58.0 |              45.0 |      unknown |     spec | 0.55 slit |     1,1 | 59181.0867630282 | 1.41291034446565 |   100.0 |    NONE |    None |     0.0 |      23 |   all |      -1 |     -1
-NR.20201128.05431.fits |  pixelflat,trace |             58.0 |              45.0 |      unknown |     spec | 0.55 slit |     1,1 |  59181.062862681 | 1.41291034446565 |   100.0 |    NONE |    None |     0.0 |      11 |   all |      -1 |     -1
-NR.20201128.05549.fits |  pixelflat,trace |             58.0 |              45.0 |      unknown |     spec | 0.55 slit |     1,1 | 59181.0642262227 | 1.41291034446565 |   100.0 |    NONE |    None |     0.0 |      12 |   all |      -1 |     -1
-NR.20201128.05665.fits |  pixelflat,trace |             58.0 |              45.0 |      unknown |     spec | 0.55 slit |     1,1 | 59181.0655779588 | 1.41291034446565 |   100.0 |    NONE |    None |     0.0 |      13 |   all |      -1 |     -1
-NR.20201128.24733.fits |         standard |        46.092875 |  14.8583611111111 |      HD19061 |     spec | 0.55 slit |     1,1 | 59181.2862718245 |  1.1546379391409 |    10.0 |    NONE |    None |     0.0 |      54 |     0 |       5 |     -1
-NR.20201128.24758.fits |         standard |        46.092875 |  14.8583611111111 |      HD19061 |     spec | 0.55 slit |     1,1 | 59181.2865551579 | 1.15348058918991 |    10.0 |    NONE |    None |     0.0 |      55 |     0 |       6 |     -1
-NR.20201128.24782.fits |         standard |        46.092875 |  14.8583611111111 |      HD19061 |     spec | 0.55 slit |     1,1 | 59181.2868384912 | 1.15244399527899 |    10.0 |    NONE |    None |     0.0 |      56 |     0 |       7 |     -1
-NR.20201128.24807.fits |         standard |        46.092875 |  14.8583611111111 |      HD19061 |     spec | 0.55 slit |     1,1 | 59181.2871277273 | 1.15125833135847 |    10.0 |    NONE |    None |     0.0 |      57 |     0 |       8 |     -1
-data end
-
